Moved listen to separte index flow
