+++ discovery-provider/src/__init__.py
@@ -258,7 +258,7 @@ def configure_celery(flask_app, celery, test_config=None):
 
     # Update celery configuration
     celery.conf.update(
-        imports=["src.tasks.index", "src.tasks.index_blacklist", "src.tasks.index_cache"],
+        imports=["src.tasks.index", "src.tasks.index_blacklist", "src.tasks.index_cache", "src.tasks.index_plays"],
         beat_schedule={
             "update_discovery_provider": {
                 "task": "update_discovery_provider",
@@ -271,6 +271,10 @@ def configure_celery(flask_app, celery, test_config=None):
             "update_cache": {
                 "task": "update_discovery_cache",
                 "schedule": timedelta(seconds=60)
+            },
+            "update_play_count": {
+                "task": "update_play_count",
+                "schedule": timedelta(seconds=10)
             }
         },
         task_serializer="json",
